backend/requirements.txt
alembic==1.13.1
blinker==1.7.0
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
cs50==9.3.4
Flask==3.0.2
Flask-Cors==4.0.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
termcolor==2.4.0
typing_extensions==4.10.0
Werkzeug==3.0.1
wheel==0.43.0
